chore: add utility function authVerification to use in protected routes
